extract ProgressMonitor class to separate file
